abstract deinflection tests
